LayoutTests/fast/canvas/canvas-filter-basics-expected.txt

This tests canavs filter setter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



context.filter should be initialized with 'none'
PASS context.filter is "none"
PASS context.filter is "invert( 75% ) sepia(1) url(#svgDropShadow)"
PASS context.filter is "invert( 75% ) sepia(1)"
PASS context.filter is "invert( 75% )"
PASS context.filter is "invert(75%)"

Empty string , null and undefined should leave context.filter unchanged
PASS context.filter is "invert(75%)"
PASS context.filter is "invert(75%)"
PASS context.filter is "invert(75%)"
PASS context.filter is "invert(75%)"

Unparsable strings should leave context.filter unchanged
PASS context.filter is "invert(75%)"
PASS context.filter is "invert(75%)"
PASS context.filter is "invert(75%)"
PASS context.filter is "invert(75%)"

context.filter = 'none'; should disable filters for the context
PASS context.filter is "none"

context.filter is part of the context state
PASS context.filter is "invert(75%)"
PASS context.filter is "sepia(1)"
PASS context.filter is "invert(75%)"
PASS successfullyParsed is true

TEST COMPLETE
